refactor(web): share timed deck layout
