+++ packages/web/src/views/Week/utils/weekTimedOverlapLayout.test.ts
@@ -2,13 +2,6 @@ import { Origin, Priorities } from "@core/constants/core.constants";
 import { type Schema_GridEvent } from "@web/common/types/web.event.types";
 import { gridEventDefaultPosition } from "@web/common/utils/event/event.util";
 import {
-  DECK_INDENT,
-  DECK_MIN_WIDTH,
-  DECK_RIGHT_RESERVE,
-  EVENT_WIDTH_MINIMUM,
-  TIMED_EVENT_COLUMN_INSET,
-} from "@web/views/Week/layout.constants";
-import {
   applyWeekTimedDeckPosition,
   createWeekTimedEventLayout,
 } from "@web/views/Week/utils/weekTimedOverlapLayout";
@@ -16,8 +9,8 @@ import {
 const createTimedEvent = (
   overrides: Partial<Schema_GridEvent> & {
     _id: string;
-    startDate: string;
     endDate: string;
+    startDate: string;
   },
 ): Schema_GridEvent => ({
   isAllDay: false,
@@ -31,183 +24,34 @@ const createTimedEvent = (
   position: { ...gridEventDefaultPosition, ...overrides.position },
 });
 
-const deckOf = (
-  items: ReturnType<typeof createWeekTimedEventLayout>,
-  id: string,
-) => items.find((item) => item.event._id === id)?.deckLayout ?? null;
-
-describe("createWeekTimedEventLayout", () => {
-  it("returns render-only deck layout for same-day overlapping events", () => {
-    const events = [
+describe("weekTimedOverlapLayout", () => {
+  it("keeps the Week layout wrapper names working", () => {
+    const items = createWeekTimedEventLayout([
       createTimedEvent({
-        _id: "early-long",
-        startDate: "2024-01-15T18:30:00.000Z",
-        endDate: "2024-01-15T20:00:00.000Z",
+        _id: "first",
+        endDate: "2026-05-20T10:00:00",
+        startDate: "2026-05-20T09:00:00",
       }),
       createTimedEvent({
-        _id: "later-short",
-        startDate: "2024-01-15T19:00:00.000Z",
-        endDate: "2024-01-15T19:45:00.000Z",
+        _id: "second",
+        endDate: "2026-05-20T10:30:00",
+        startDate: "2026-05-20T09:30:00",
       }),
-    ];
+    ]);
 
-    const laid = createWeekTimedEventLayout(events);
-
-    expect(deckOf(laid, "early-long")).toEqual({ order: 0, groupSize: 2 });
-    expect(deckOf(laid, "later-short")).toEqual({ order: 1, groupSize: 2 });
-    expect(laid[0].event).toBe(events[0]);
-    expect(laid[1].event).toBe(events[1]);
+    expect(items.map((item) => item.deckLayout)).toEqual([
+      { groupSize: 2, order: 0 },
+      { groupSize: 2, order: 1 },
+    ]);
   });
 
-  it("orders same-start events longest-first", () => {
-    const events = [
-      createTimedEvent({
-        _id: "short",
-        startDate: "2024-01-15T18:00:00.000Z",
-        endDate: "2024-01-15T18:30:00.000Z",
-      }),
-      createTimedEvent({
-        _id: "long",
-        startDate: "2024-01-15T18:00:00.000Z",
-        endDate: "2024-01-15T20:00:00.000Z",
-      }),
-    ];
-
-    const laid = createWeekTimedEventLayout(events);
-
-    expect(deckOf(laid, "long")).toEqual({ order: 0, groupSize: 2 });
-    expect(deckOf(laid, "short")).toEqual({ order: 1, groupSize: 2 });
-  });
-
-  it("treats a transitive overlap chain as one group", () => {
-    const events = [
-      createTimedEvent({
-        _id: "a",
-        startDate: "2024-01-15T10:00:00.000Z",
-        endDate: "2024-01-15T11:00:00.000Z",
-      }),
-      createTimedEvent({
-        _id: "b",
-        startDate: "2024-01-15T10:30:00.000Z",
-        endDate: "2024-01-15T11:30:00.000Z",
-      }),
-      createTimedEvent({
-        _id: "c",
-        startDate: "2024-01-15T11:00:00.000Z",
-        endDate: "2024-01-15T12:00:00.000Z",
-      }),
-    ];
-
-    const laid = createWeekTimedEventLayout(events);
-
-    expect(deckOf(laid, "a")).toEqual({ order: 0, groupSize: 3 });
-    expect(deckOf(laid, "b")).toEqual({ order: 1, groupSize: 3 });
-    expect(deckOf(laid, "c")).toEqual({ order: 2, groupSize: 3 });
-  });
-
-  it("leaves non-overlapping events without deck layout", () => {
-    const events = [
-      createTimedEvent({
-        _id: "morning",
-        startDate: "2024-01-15T09:00:00.000Z",
-        endDate: "2024-01-15T10:00:00.000Z",
-      }),
-      createTimedEvent({
-        _id: "afternoon",
-        startDate: "2024-01-15T14:00:00.000Z",
-        endDate: "2024-01-15T15:00:00.000Z",
-      }),
-    ];
-
-    const laid = createWeekTimedEventLayout(events);
-
-    expect(deckOf(laid, "morning")).toBeNull();
-    expect(deckOf(laid, "afternoon")).toBeNull();
-  });
-});
-
-describe("applyWeekTimedDeckPosition", () => {
-  const basePosition = { height: 60, left: 20, top: 30, width: 180 };
-
-  it("shrinks deck width by the reserve and group indents", () => {
-    const deck = applyWeekTimedDeckPosition(basePosition, {
-      order: 0,
-      groupSize: 3,
-    });
-
-    expect(deck.width).toBe(
-      basePosition.width - DECK_RIGHT_RESERVE - 2 * DECK_INDENT,
-    );
-  });
-
-  it("indents left by order and sets zIndex", () => {
-    const back = applyWeekTimedDeckPosition(basePosition, {
-      order: 0,
-      groupSize: 3,
-    });
-    const front = applyWeekTimedDeckPosition(basePosition, {
-      order: 2,
-      groupSize: 3,
-    });
-
-    expect(front.left - back.left).toBe(2 * DECK_INDENT);
-    expect(back.zIndex).toBe(1);
-    expect(front.zIndex).toBe(3);
-    expect(front.width).toBe(back.width);
-  });
-
-  it("leaves a wider visible rail for cards behind a same-time overlap", () => {
-    const back = applyWeekTimedDeckPosition(basePosition, {
-      order: 0,
-      groupSize: 2,
-    });
-    const front = applyWeekTimedDeckPosition(basePosition, {
-      order: 1,
-      groupSize: 2,
-    });
-
-    expect(front.left - back.left).toBe(DECK_INDENT);
-  });
-
-  it("floors deck width for dense groups", () => {
-    const deck = applyWeekTimedDeckPosition(
-      { ...basePosition, width: 150 - TIMED_EVENT_COLUMN_INSET * 2 },
-      { order: 0, groupSize: 8 },
+  it("keeps the Week deck-position wrapper name working", () => {
+    const position = applyWeekTimedDeckPosition(
+      { height: 60, left: 20, top: 30, width: 180 },
+      { groupSize: 2, order: 1 },
     );
 
-    expect(deck.width).toBe(DECK_MIN_WIDTH);
-  });
-
-  it("keeps narrow-column decks inside the usable column", () => {
-    const narrowBase = {
-      ...basePosition,
-      width: EVENT_WIDTH_MINIMUM - TIMED_EVENT_COLUMN_INSET * 2,
-    };
-    const back = applyWeekTimedDeckPosition(narrowBase, {
-      order: 0,
-      groupSize: 2,
-    });
-    const front = applyWeekTimedDeckPosition(narrowBase, {
-      order: 1,
-      groupSize: 2,
-    });
-
-    expect(back.width).toBe(narrowBase.width - DECK_INDENT);
-    expect(front.width).toBe(back.width);
-    expect(front.left + front.width).toBe(narrowBase.left + narrowBase.width);
-  });
-
-  it("keeps dense narrow-column decks inside the usable column", () => {
-    const narrowBase = {
-      ...basePosition,
-      width: EVENT_WIDTH_MINIMUM - TIMED_EVENT_COLUMN_INSET * 2,
-    };
-    const front = applyWeekTimedDeckPosition(narrowBase, {
-      order: 7,
-      groupSize: 8,
-    });
-
-    expect(front.width).toBeGreaterThan(0);
-    expect(front.left + front.width).toBe(narrowBase.left + narrowBase.width);
+    expect(position.left).toBeGreaterThan(20);
+    expect(position.zIndex).toBe(2);
   });
 });
